Updated botapi

+++ src/no/northcode/jens/intranetsek2tg/DataHandler.java
@@ -13,8 +13,8 @@ public class DataHandler {
 
 	private static final String CONFIG_FILE = "config.json";
 	
-	public static HashMap<Integer, UserData> loadUsers() {
-		HashMap<Integer, UserData> users = new HashMap<Integer, UserData>();
+	public static HashMap<Long, UserData> loadUsers() {
+		HashMap<Long, UserData> users = new HashMap<Long, UserData>();
 		
 		try {
 			JSONParser parser = new JSONParser();
@@ -27,7 +27,7 @@ public class DataHandler {
 				u.school = (String) juser.get("school");
 				u.password = (String) juser.get("password");
 				u.state = UserState.valueOf((String) juser.get("state"));
-				users.put(((Long) juser.get("userid")).intValue(), u); 
+				users.put((Long) juser.get("userid"), u); 
 			}
 		} catch (Exception ex) { 
 			System.out.println("Config file not found!");
@@ -38,7 +38,7 @@ public class DataHandler {
 		return users;
 	}
 	
-	public static HashMap<String, GroupData> loadGroups(HashMap<Integer, UserData> users) {
+	public static HashMap<String, GroupData> loadGroups(HashMap<Long, UserData> users) {
 		HashMap<String, GroupData> groups = new HashMap<String, GroupData>();
 		
 		try {
@@ -58,11 +58,11 @@ public class DataHandler {
 	}
 	
 	@SuppressWarnings("unchecked")
-	public static void saveData(HashMap<Integer, UserData> users, HashMap<String, GroupData> groups) throws IOException {
+	public static void saveData(HashMap<Long, UserData> users, HashMap<String, GroupData> groups) throws IOException {
 		JSONObject root = new JSONObject();
 		JSONArray userlist = new JSONArray();
 		// Users
-		for(Integer userid : users.keySet()) {
+		for(Long userid : users.keySet()) {
 			UserData user = users.get(userid);
 			JSONObject udata = new JSONObject();
 			udata.put("userid", userid);
